Added convenience function for solve_symmetric. Moved the Op lifter for Images2Neibs in opt.py such that it is included in the optimisation passes. Closes #6158
